+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.1.5
+
+### Enhancements
+- Add 'Copy Breadcrumbs' button to the Rocket Documentation pages.  
+
+### Tweaks
+- Disable extension update Windows notification in Chrome/Edge.  
+
 # 4.1.4
 
 ### Fixes